pyproject.toml
[project]
name = "pdf-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "nltk>=3.9.4",
    "numpy>=2.4.4",
    "pymupdf>=1.27.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
]
